requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
gym==0.23.1
gym-notices==0.0.8
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
pygame==2.3.0
pyparsing==3.0.9
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
tzdata==2023.3
urllib3==1.26.15
